Remove explicit require of rspec to fix builds
Not sure why this works, or is even possible? But I can reproduce locally that
all non-1.8 builds fail when using a bundler cache unless this line is removed.

My guess is that, when using the --path option to bundler, all gems are being
implicitly required, and requiring rspec a second time is somehow causing issues,
but that doesn't make much sense to me.