Update bundler before Travis CI builds
Builds are broken because the `bundler` version that Travis CI
automatically uses at this time (`1.7.6`) is coming across an
error that was fixed in a later version/release.

Having an outdated version of `bundler` in the Travis CI builds is an
ongoing issue and this is currently the easiest work-around.

See:

* bundler/bundler#3558
* travis-ci/travis-ci#3531

(commit msg taken from: https://github.com/thoughtbot/factory_girl/commit/3c0547fa508d212f79050301d5622434e4731c48)