Call #instance_exec on DSL object instead of proxy
when the DSL object is the same object as the block context object.

This short-circuits prior to performing setup (like copying ivars
from the block context to the proxy) that is unnecessary when the block
context and DSL object "match".

Add test that runs non-mutative code on frozen DSL object to check that
no FrozenError is raised, as it would have been if a
FallbackContextProxy tried to dynamically add methods to the DSL object.