Move Gemfile CI-only dependencies to test group
We want to be careful to avoid accidentally bringing
in runtime dependencies to projects which use Docile.

In this PR, we clarify explicitly that the CI-only dependency
on simplecov-cobertura, which is the formatt used by CodeCov,
is a test dependency, and not a runtime dependency.

Without this change, it may be possible that projects which use
Docile, when running in an environment with CI=true, would
accidentally bring in this test-only dependency, when that isn't
intended.

  - Remove unnecessary quoting of '3.1' in .github/workflows/main.yml